sns/requirements.txt
certifi==2023.7.22
charset-normalizer==3.3.1
contourpy==1.1.1
cycler==0.12.1
fonttools==4.43.1
idna==3.4
iniconfig==2.0.0
kiwisolver==1.4.5
llvmlite==0.41.1
matplotlib==3.8.0
networkx==3.1
ns.py==0.3.7
numba==0.58.1
numpy==1.26.0
packaging==23.2
Pillow==10.1.0
pluggy==1.3.0
pyparsing==3.1.1
pytest==7.4.2
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
scipy==1.11.3
simpy==4.0.2
six==1.16.0
urllib3==2.0.7
